requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0
PyYAML>=6.0